rename : style 파일 확장자 변경
