+++ src/style/skeletonUI/skeletonPage/DetailSkeleton.tsx
