+++ src/style/skeletonUI/skeletonPage/MainSkeleton.tsx
