+++ src/style/skeletonUI/skeletonPage/SearchSkeleton.tsx
